+++ game/board.py
@@ -4,7 +4,7 @@ from piezas.horse import Horse
 from piezas.pawn import Pawn
 from piezas.bishop import Bishop
 from piezas.queen import Queen
-
+from game.exceptions import OutOfBoard
 
 
 class Board:
@@ -16,42 +16,40 @@ class Board:
                 col.append(None)
             self.__positions__.append(col)
         if not for_test:
-
-
-        #POSICIONES TORRES
+            #POSICIONES TORRES
             self.__positions__[0][0] = Rook("BLACK") # Black
             self.__positions__[0][7] = Rook("BLACK") # Black
-            self.__positions__[7][7] = Rook("WHITE") # White
-            self.__positions__[7][0] = Rook("WHITE") # White
-        #PISICIONES REY
+            self.__positions__[7][0] = Rook("WHITE") 
+            self.__positions__[7][7] = Rook("WHITE") # White# White
+            #PISICIONES REY
             self.__positions__[0][4] = King("BLACK") 
             self.__positions__[7][4] = King("WHITE") 
-
+            #POSICIONES CABALLOS
             self.__positions__[0][1] = Horse("BLACK") 
             self.__positions__[0][6] = Horse("BLACK") 
             self.__positions__[7][1] = Horse("WHITE") 
             self.__positions__[7][6] = Horse("WHITE") 
-
             #PISICION ALFIL
             self.__positions__[0][2] = Bishop("BLACK") 
             self.__positions__[0][5] = Bishop("BLACK") 
             self.__positions__[7][2] = Bishop("WHITE") 
             self.__positions__[7][5] = Bishop("WHITE") 
-
             #POSICIONES REINA    
             self.__positions__[0][3] = Queen("BLACK") 
             self.__positions__[7][3] = Queen("WHITE")
-
-        #POSICIONES PEONES
+            #POSICIONES PEONES
             for col in range(8):
                 self.__positions__[6][col] = Pawn("White")
                 self.__positions__[1][col] = Pawn("Black")
 
+
         
 
     def get_piece(self, row, col):
+        if 0 <= row < 8 and 0 <= col < 8:
             return self.__positions__[row][col]
-    
+        else:
+            raise OutOfBoard("Coordinates out of range")
 
 
 
@@ -65,19 +63,28 @@ class Board:
                     board_str += " "
             board_str += "\n"
         return board_str
-
+    
+ 
+   
    
     def set_piece(self, row, col, piece):
         self.__positions__[row][col] = piece
-
-    
-
 
     
     def move(self, from_row, from_col, to_row, to_col):
         origin = self.get_piece(from_row, from_col)
         self.set_piece(to_row, to_col, origin)
         self.set_piece(from_row, from_col, None)
+
+    #OBTENER MOVIMIENTOS VALIDOS
+    def ver_movimientos_validos(self, row, col):
+        piece = self.get_piece(row, col)
+        if piece is not None:
+            return piece.control(row, col, self)
+        else:
+            return []
+        
+