+++ piezas/horse.py
@@ -1,28 +1,14 @@
 
 from piezas.piece import Piece
 class Horse(Piece):
-    black_str ="♘" #knight black
-    white_str ="♞"  #knight white
+    black_str ="♘" 
+    white_str ="♞"  
     def __init__(self, color):
         super().__init__(color)
 
-    def pieza_del_mismo_color(self, row, col, board):
-        #Esta funcion sirve para ver si hay otra pieza del mismo color
-        pieza = board.get_piece(row,col)
-        if pieza is not None and pieza.__color__ == self.__color__:
-            return True
-        else:
-            return False
-        
-    def captura(self, row, col, board):
-        #Verifica si puede capturar 
-        pieza = board.get_piece(row,col)
-        if pieza is not None and pieza.__color__ != self.__color__:
-            return True
-        else:
-            return False
 
-    def horse_moves(self,row, col,board):
+
+    def movimientos_caballo(self,row, col,board):
         moves = []
     # Lista de movimientos posibles en forma de L
         possible_moves = [