+++ piezas/king.py
@@ -5,24 +5,10 @@ class King(Piece):
     def __init__(self, color):
         super().__init__(color)
 
-    def pieza_del_mismo_color(self, row, col, board):
-        #Esta funcion sirve para ver si hay otra pieza del mismo color
-        pieza = board.get_piece(row,col)
-        if pieza is not None and pieza.__color__ == self.__color__:
-            return True
-        else:
-            return False
+
         
-    def captura(self, row, col, board):
-        #Verifica si la torre puede capturar 
-        pieza = board.get_piece(row,col)
-        if pieza is not None and pieza.__color__ != self.__color__:
-            return True
-        else:
-            return False
-        
-   
-    def valid_move_king(self, from_row, from_col, board):
+    #VERIFICA LOS MOVIMIENTOS DIAGONALES,HORIZONTALES Y VERTICALES, SOLO AVAZA UNA CASILLA 
+    def movimientos_rey(self, from_row, from_col, board):
         nuevos_movimientos = []
         direcciones = [(-1, -1), (-1, 0), (-1, 1), (0, -1), (0, 1), (1, -1), (1, 0), (1, 1)]