+++ piezas/pawn.py
@@ -3,6 +3,8 @@ from piezas.piece import Piece
 
 
 class Pawn(Piece):
+    black_str = "♙"
+    white_str = "♟"
     def __init__(self, color):
         super().__init__(color)